Crate bee_tangle

A crate that contains foundational building blocks for the IOTA Tangle.

Re-exports

pub use ms_tangle::MsTangle;

pub use urts::BELOW_MAX_DEPTH;

pub use worker::TangleWorker;

Modules

config

flags

metadata

ms_tangle

pruning

storage

tangle

traversal
Collection of Tangle traversal functions.

urts

vertex

worker

Structs

MessageRef
A thread-safe reference to a Message.

Tangle
A foundational, thread-safe graph datastructure to represent the IOTA Tangle.

Traits

Hooks
A trait used to provide hooks for a tangle. The tangle acts as an in-memory cache and will use hooks to extend its effective volume. When an entry doesn't exist in the tangle cache and needs fetching, or when an entry gets inserted, the tangle will call out to the hooks in order to fulfil these actions.

Functions

init