Struct bee_tangle::ms_tangle::MsTangle

pub struct MsTangle<B> {
    pub(crate) inner: Tangle<MessageMetadata, StorageHooks<B>>,
    pub(crate) milestones: Mutex<HashMap<MilestoneIndex, Milestone>>,
    pub(crate) solid_entry_points: Mutex<HashMap<SolidEntryPoint, MilestoneIndex>>,
    latest_milestone_index: AtomicU32,
    latest_solid_milestone_index: AtomicU32,
    snapshot_index: AtomicU32,
    pruning_index: AtomicU32,
    entry_point_index: AtomicU32,
    tip_pool: Mutex<UrtsTipPool>,
}

Milestone-based Tangle.

Fields
inner: Tangle<MessageMetadata, StorageHooks<B>>milestones: Mutex<HashMap<MilestoneIndex, Milestone>>solid_entry_points: Mutex<HashMap<SolidEntryPoint, MilestoneIndex>>latest_milestone_index: AtomicU32latest_solid_milestone_index: AtomicU32snapshot_index: AtomicU32pruning_index: AtomicU32entry_point_index: AtomicU32tip_pool: Mutex<UrtsTipPool>
Implementations

impl<B: StorageBackend> MsTangle<B>

pub fn new(storage: ResourceHandle<B>) -> Self

pub async fn shutdown(self)

pub async fn insert<'_>(
&'_ self,
message: Message,
hash: MessageId,
metadata: MessageMetadata
) -> Option<MessageRef>

pub async fn add_milestone<'_>(
&'_ self,
idx: MilestoneIndex,
milestone: Milestone
)

pub async fn remove_milestone<'_>(&'_ self, index: MilestoneIndex)

async fn pull_milestone<'_>(&'_ self, idx: MilestoneIndex) -> Option<MessageId>

pub async fn get_milestone<'_>(
&'_ self,
index: MilestoneIndex
) -> Option<MessageRef>

pub async fn get_milestone_message_id<'_>(
&'_ self,
index: MilestoneIndex
) -> Option<MessageId>

pub async fn contains_milestone<'_>(&'_ self, idx: MilestoneIndex) -> bool

pub fn get_latest_milestone_index(&self) -> MilestoneIndex

pub fn update_latest_milestone_index(&self, new_index: MilestoneIndex)

pub fn get_latest_solid_milestone_index(&self) -> MilestoneIndex

pub fn update_latest_solid_milestone_index(&self, new_index: MilestoneIndex)

pub fn get_snapshot_index(&self) -> MilestoneIndex

pub fn update_snapshot_index(&self, new_index: MilestoneIndex)

pub fn get_pruning_index(&self) -> MilestoneIndex

pub fn update_pruning_index(&self, new_index: MilestoneIndex)

pub fn get_entry_point_index(&self) -> MilestoneIndex

pub fn update_entry_point_index(&self, new_index: MilestoneIndex)

pub fn is_synced(&self) -> bool

pub fn is_synced_threshold(&self, threshold: u32) -> bool

pub async fn get_solid_entry_point_index<'_, '_>(
&'_ self,
sep: &'_ SolidEntryPoint
) -> Option<MilestoneIndex>

pub async fn add_solid_entry_point<'_>(
&'_ self,
sep: SolidEntryPoint,
index: MilestoneIndex
)

pub async fn remove_solid_entry_point<'_, '_>(
&'_ self,
sep: &'_ SolidEntryPoint
)

Removes sep from the set of solid entry points.

pub async fn clear_solid_entry_points<'_>(&'_ self)

pub async fn is_solid_entry_point<'_, '_>(&'_ self, id: &'_ MessageId) -> bool

Returns whether the message associated with sep is a solid entry point.

pub async fn is_solid_message<'_, '_>(&'_ self, id: &'_ MessageId) -> bool

Returns whether the message associated with id is deemed solid.

pub async fn is_solid_message_maybe<'_, '_>(&'_ self, id: &'_ MessageId) -> bool

pub async fn otrsi<'_, '_>(&'_ self, id: &'_ MessageId) -> Option<IndexId>

pub async fn ytrsi<'_, '_>(&'_ self, id: &'_ MessageId) -> Option<IndexId>

pub async fn insert_tip<'_>(
&'_ self,
message_id: MessageId,
parents: Vec<MessageId>
)

pub async fn update_tip_scores<'_>(&'_ self)

pub async fn get_messages_to_approve<'_>(&'_ self) -> Option<Vec<MessageId>>

pub async fn reduce_tips<'_>(&'_ self)

pub async fn non_lazy_tips_num<'_>(&'_ self) -> usize

pub async fn is_healthy<'_>(&'_ self) -> bool