Struct bee_tangle::Tangle

pub struct Tangle<T, H = NullHooks<T>> where
    T: Clone,  {
    vertices: TRwLock<HashMap<MessageId, Vertex<T>>>,
    children: TRwLock<HashMap<MessageId, (HashSet<MessageId>, bool)>>,
    pub(crate) cache_counter: AtomicU64,
    pub(crate) cache_queue: Mutex<LruCache<MessageId, u64, DefaultHashBuilder>>,
    pub(crate) hooks: H,
}

A foundational, thread-safe graph datastructure to represent the IOTA Tangle.

Fields
vertices: TRwLock<HashMap<MessageId, Vertex<T>>>children: TRwLock<HashMap<MessageId, (HashSet<MessageId>, bool)>>cache_counter: AtomicU64cache_queue: Mutex<LruCache<MessageId, u64, DefaultHashBuilder>>hooks: H
Implementations

impl<T, H: Hooks<T>> Tangle<T, H> where
T: Clone,

pub fn new(hooks: H) -> Self

Creates a new Tangle.

pub fn with_capacity(self, cap: usize) -> Self

Create a new tangle with the given capacity.

pub fn hooks(&self) -> &H

Return a reference to the storage hooks used by this tangle.

async fn insert_inner<'_>(
&'_ self,
message_id: MessageId,
message: Message,
metadata: T
) -> Option<MessageRef>

pub async fn insert<'_>(
&'_ self,
message_id: MessageId,
message: Message,
metadata: T
) -> Option<MessageRef>

Inserts a message, and returns a thread-safe reference to it in case it didn't already exist.

async fn add_child_inner<'_>(&'_ self, parent: MessageId, child: MessageId)

async fn get_inner<'_, '_, '_>(
&'_ self,
message_id: &'_ MessageId
) -> Option<impl Deref<Target = Vertex<T>> + '_>

pub async fn get<'_, '_>(
&'_ self,
message_id: &'_ MessageId
) -> Option<MessageRef>

Get the data of a vertex associated with the given message_id.

async fn contains_inner<'_, '_>(&'_ self, message_id: &'_ MessageId) -> bool

pub async fn contains<'_, '_>(&'_ self, message_id: &'_ MessageId) -> bool

Returns whether the message is stored in the Tangle.

pub async fn get_metadata<'_, '_>(
&'_ self,
message_id: &'_ MessageId
) -> Option<T>

Get the metadata of a vertex associated with the given message_id.

pub async fn get_metadata_maybe<'_, '_>(
&'_ self,
message_id: &'_ MessageId
) -> Option<T>

Get the metadata of a vertex associated with the given message_id, if it's in the cahce.

pub async fn get_vertex<'_, '_, '_>(
&'_ self,
message_id: &'_ MessageId
) -> Option<impl Deref<Target = Vertex<T>> + '_>

Get the metadata of a vertex associated with the given message_id.

pub async fn set_metadata<'_, '_>(
&'_ self,
message_id: &'_ MessageId,
metadata: T
)

Updates the metadata of a particular vertex.

pub async fn update_metadata<R, Update, '_, '_>(
&'_ self,
message_id: &'_ MessageId,
update: Update
) -> Option<R> where
Update: FnMut(&mut T) -> R,

Updates the metadata of a vertex.

pub async fn len<'_>(&'_ self) -> usize

Returns the number of messages in the Tangle.

pub async fn is_empty<'_>(&'_ self) -> bool

Checks if the tangle is empty.

async fn children_inner<'_, '_, '_>(
&'_ self,
message_id: &'_ MessageId
) -> Option<impl Deref<Target = HashSet<MessageId>> + '_>

pub async fn get_children<'_, '_>(
&'_ self,
message_id: &'_ MessageId
) -> Option<HashSet<MessageId>>

Returns the children of a vertex, if we know about them.

pub async fn num_children<'_, '_>(&'_ self, message_id: &'_ MessageId) -> usize

Returns the number of children of a vertex.

async fn pull_message<'_, '_>(&'_ self, message_id: &'_ MessageId) -> bool

fn generate_cache_index(&self) -> u64

async fn perform_eviction<'_>(&'_ self)