Trait bee_tangle::tangle::Hooks

pub trait Hooks<T> {
    type Error: Debug;
#[must_use]    fn get<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        message_id: &'life1 MessageId
    ) -> Pin<Box<dyn Future<Output = Result<Option<(Message, T)>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn insert<'life0, 'async_trait>(
        &'life0 self, 
        message_id: MessageId, 
        tx: Message, 
        metadata: T
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn fetch_approvers<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        message_id: &'life1 MessageId
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<MessageId>>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn insert_approver<'life0, 'async_trait>(
        &'life0 self, 
        message_id: MessageId, 
        approver: MessageId
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn update_approvers<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        message_id: MessageId, 
        approvers: &'life1 Vec<MessageId>
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

A trait used to provide hooks for a tangle. The tangle acts as an in-memory cache and will use hooks to extend its effective volume. When an entry doesn't exist in the tangle cache and needs fetching, or when an entry gets inserted, the tangle will call out to the hooks in order to fulfil these actions.

Associated Types

type Error: Debug

An error generated by these hooks.

Loading content...
Required methods

#[must_use]fn get<'life0, 'life1, 'async_trait>(
&'life0 self,
message_id: &'life1 MessageId
) -> Pin<Box<dyn Future<Output = Result<Option<(Message, T)>, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Fetch a message from some external storage medium.

#[must_use]fn insert<'life0, 'async_trait>(
&'life0 self,
message_id: MessageId,
tx: Message,
metadata: T
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Insert a message into some external storage medium.

#[must_use]fn fetch_approvers<'life0, 'life1, 'async_trait>(
&'life0 self,
message_id: &'life1 MessageId
) -> Pin<Box<dyn Future<Output = Result<Option<Vec<MessageId>>, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Fetch the approvers list for a given message.

#[must_use]fn insert_approver<'life0, 'async_trait>(
&'life0 self,
message_id: MessageId,
approver: MessageId
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Insert a new approver for a given message.

#[must_use]fn update_approvers<'life0, 'life1, 'async_trait>(
&'life0 self,
message_id: MessageId,
approvers: &'life1 Vec<MessageId>
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Update the approvers list for a given message.

Loading content...